packages/agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strategic-research-copilot"
version = "0.1.0"
description = "AI research analyst with LangGraph orchestration, knowledge graphs, and MCP integrations"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Khushaal Chaudhary"}
]
keywords = ["langgraph", "neo4j", "mcp", "rag", "knowledge-graph", "research", "agent"]

dependencies = [
    # LangChain ecosystem
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    
    # Neo4j
    "langchain-neo4j>=0.2.0",
    "neo4j>=5.0.0",
    
    # LLM Providers
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.0",
    "langchain-ollama>=0.2.0",         # For local Ollama
    "langchain-openai>=0.2.0",         # For OpenAI
    
    # Embeddings & Search
    "sentence-transformers>=3.0.0",
    
    # Graph analysis
    "networkx>=3.0",
    "python-louvain>=0.16",
    
    # Document processing
    "langchain-text-splitters>=0.3.0",
    
    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # MCP Client
    "mcp>=1.0.0",
    
    # HTTP client for APIs
    "httpx>=0.27.0",
    
    # Web Search (Tavily - AI-powered search)
    "tavily-python>=0.5.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "tqdm>=4.0.0",
    
    # CLI
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]

[project.scripts]
copilot = "copilot.interfaces.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/copilot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"